requirements.txt
requests
beautifulsoup4
schedule
Flask
python-dotenv